PARSER-03: extract asm case value helpers
